pyproject.toml
[tool.poetry]
name = "lms-system-drf"
version = "0.1.0"
description = ""
authors = ["Uliana Semyonova <i@ulianasemyonova.ru>"]
readme = "README.md"
packages = [{include = "lms_system_drf"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
djangorestframework = "^3.14.0"
pillow = "^10.0.1"
psycopg2 = "^2.9.8"
python-dotenv = "^1.0.0"
django-filter = "^23.3"
djangorestframework-simplejwt = "^5.3.0"
coverage = "^7.3.2"
drf-yasg = "^1.21.7"
stripe = "^7.0.0"
django-crontab = "^0.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
